rename : Contents관련 태그들을 세부 컴포넌트로 분할해서 정리
