+++ src/main/front/src/components/ContentsInfo/InfoComments.js
@@ -1,0 +1,25 @@
+import React, { useState } from "react";
+
+import ContentModal from "../Modal/ContentModal";
+
+import styles from "../../styles/Contents.module.css";
+
+function InfoComments(props) {
+  const [openDetail, setOpenDetail] = useState(false);
+  const handleOpenDetail = () => {
+    setOpenDetail(true);
+  };
+  const handleCloseDetail = () => {
+    setOpenDetail(false);
+  };
+  return (
+    <>
+      <button className={styles.allComments} onClick={handleOpenDetail}>
+        {props.comments} view all comments
+      </button>
+      {openDetail == true ? <ContentModal onClick={handleCloseDetail} /> : null}
+    </>
+  );
+}
+
+export default InfoComments;
