+++ src/main/front/src/pages/main/Main.js
@@ -1,6 +1,6 @@
 import React from "react";
 
-import Contents from "../../components/Contents";
+import Contents from "../../components/Contents/Contents";
 import styles from "../../styles/Contents.module.css";
 export default function Main() {
   const datas = [
@@ -92,7 +92,7 @@ export default function Main() {
       userId={data.feed_id}
       liked={data.total_like_number}
       comments={data.comment.length}
-    ></Contents>
+    />
   ));
 
   return (
